Moved VMI get ips function to global
