+++ src/utils/resources/vmi/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './affinity';
 export * from './cpu';
 export * from './flavor';
+export * from './ips';
 export * from './nodeSelector';
