+++ src/views/virtualmachines/list/components/FirstItemListPopover/FirstItemListPopover.tsx
@@ -1,22 +1,11 @@
 import * as React from 'react';
 
-import { V1VirtualMachineInstance } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import { Button, ButtonVariant, Popover, PopoverPosition } from '@patternfly/react-core';
 
 type FirstItemListPopoverProps = {
   items: string[];
   headerContent?: React.ReactNode | string;
   className?: string;
-};
-
-export const getVMIIPAddresses = (vmi: V1VirtualMachineInstance): string[] => {
-  const namedInterfaces = vmi?.status?.interfaces?.filter((iface) => !!iface.name) || [];
-  const ipAddresses = namedInterfaces?.flatMap((iface) => [
-    iface?.ipAddress,
-    ...(iface?.ipAddresses || []),
-  ]);
-  const trimmedIPAddresses = ipAddresses?.map((ip) => ip.trim())?.filter((ip) => ip.length > 0);
-  return [...new Set(trimmedIPAddresses)];
 };
 
 const FirstItemListPopover: React.FC<FirstItemListPopoverProps> = ({
